feat(types): an item can now be a list
An item (e.g. a single tensor, a np.array or a tuple/dict containing data and
labels) can now be a list. This is needed in PyTorch when getting a batch of
a data loader: for example `next(iter(dataloader))` returns a list of two
torch.Tensor [images, labels]. TensorFlow returns a tuple instead.