feat: add logits to feature extractors outputs
At feature extractor construction, the output_layers_id are used to return
features. The model output, i.e the logits, are now also returned.
As a consequence, predict_tensor() always returns a list of features + logits,
even for a single tensor.

Note that, for PyTorch feature extractor, a new member `self._hook_layers_id`
is introduced: it is simply the output_layers_id + the last layer id (-1) for
the model logits. This new member variable is used to register the hooks in
prepare_extractor() and to return the features+logits in predict_tensor().